need more improvment on the algo

+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: abdnasse <abdnasse@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/27 17:08:22 by abdnasse          #+#    #+#             */
-/*   Updated: 2025/02/06 19:26:52 by abdnasse         ###   ########.fr       */
+/*   Updated: 2025/02/06 20:31:41 by abdnasse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 typedef struct s_list
 {
 	int	content;
-	size_t	index;
+	int	index;
 	struct s_list	*next;
 }					t_list;
 void	sort_algo(t_list **a, t_list **b, int size);
